fix: Files have been moved to the desired directory
